Migrate Component to Function Component
