+++ opencti-platform/opencti-front/src/private/components/nav/TopMenuEvents.tsx
@@ -1,14 +1,13 @@
-import React, { Component } from 'react';
-import * as PropTypes from 'prop-types';
-import { withRouter, Link } from 'react-router-dom';
-import { compose } from 'ramda';
-import withStyles from '@mui/styles/withStyles';
+import React from 'react';
+import { Link, useLocation } from 'react-router-dom';
 import Button from '@mui/material/Button';
-import { WifiTetheringOutlined, VisibilityOutlined } from '@mui/icons-material';
+import { VisibilityOutlined, WifiTetheringOutlined } from '@mui/icons-material';
 import { Fire } from 'mdi-material-ui';
-import inject18n from '../../../components/i18n';
+import makeStyles from '@mui/styles/makeStyles';
+import { Theme } from '@mui/material/styles/createTheme';
+import { useFormatter } from '../../../components/i18n';
 
-const styles = (theme) => ({
+const useStyles = makeStyles<Theme>((theme) => ({
   button: {
     marginRight: theme.spacing(2),
     padding: '0 5px 0 5px',
@@ -19,12 +18,14 @@ const styles = (theme) => ({
   icon: {
     marginRight: theme.spacing(1),
   },
-});
+}));
 
-class TopMenuThreats extends Component {
-  render() {
-    const { t, location, classes } = this.props;
-    return (
+const TopMenuEvents = () => {
+  const classes = useStyles();
+  const location = useLocation();
+  const { t } = useFormatter();
+
+  return (
       <div>
         <Button
           component={Link}
@@ -84,19 +85,7 @@ class TopMenuThreats extends Component {
           {t('Observed datas')}
         </Button>
       </div>
-    );
-  }
-}
-
-TopMenuThreats.propTypes = {
-  classes: PropTypes.object,
-  location: PropTypes.object,
-  t: PropTypes.func,
-  history: PropTypes.object,
+  );
 };
 
-export default compose(
-  inject18n,
-  withRouter,
-  withStyles(styles),
-)(TopMenuThreats);
+export default TopMenuEvents;
